src/app4/requirements.txt
# Requirements for src/app4/app4.py
python-dotenv
azure-identity
azure-ai-projects
azure-ai-agents
azure-monitor-opentelemetry
opentelemetry-api